feat: squash 18 alembic migrations into single initial_schema migration
Squashed migrations:
- 883740308b56 through 3d846b7d1114 (18 migrations total)

Changes:
- Deleted 18 individual migration files
- Created single squashed migration (237745010854_initial_schema.py)
- Preserved all schema elements:
  * 19 tables with complete schema
  * 5 enum types (all with create_constraint=True)
  * 3 partial unique indexes (WHERE deleted_at IS NULL)
  * 10+ column comments
  * 1 CHECK constraint (notification_preferences)
  * 3 server defaults (users.auth_provider, lines.mode, line_disruption_state_logs timestamps)
  * Data population: 9 rows in alert_disabled_severities

Manual additions (not autogenerated):
- server_default for lines.mode ('tube')
- server_defaults for line_disruption_state_logs.created_at/updated_at (now())
- Data population for alert_disabled_severities (Good Service for all modes)
- downgrade() raises NotImplementedError

Issue: #238
