+++ backend/alembic/versions/237745010854_initial_schema.py
@@ -1,0 +1,527 @@
+"""initial_schema
+
+Revision ID: 237745010854
+Revises:
+Create Date: 2025-11-22 22:04:40.213028
+
+"""
+
+from collections.abc import Sequence
+
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "237745010854"
+down_revision: str | Sequence[str] | None = None
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:  # noqa: PLR0915 - Squashed migration has many statements
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "alert_disabled_severities",
+        sa.Column(
+            "mode_id",
+            sa.String(length=50),
+            nullable=False,
+            comment="Transport mode (e.g., 'tube', 'dlr', 'overground')",
+        ),
+        sa.Column("severity_level", sa.Integer(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("mode_id", "severity_level", name="uq_alert_disabled_severity_mode_level"),
+    )
+    op.create_index(
+        op.f("ix_alert_disabled_severities_mode_id"), "alert_disabled_severities", ["mode_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_alert_disabled_severities_severity_level"),
+        "alert_disabled_severities",
+        ["severity_level"],
+        unique=False,
+    )
+    op.create_table(
+        "disruption_categories",
+        sa.Column("category_name", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.String(length=500), nullable=True),
+        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_disruption_categories_category_name"), "disruption_categories", ["category_name"], unique=True
+    )
+    op.create_table(
+        "line_disruption_state_logs",
+        sa.Column(
+            "line_id", sa.String(length=50), nullable=False, comment="TfL line ID (e.g., 'bakerloo', 'victoria')"
+        ),
+        sa.Column(
+            "status_severity_description",
+            sa.String(length=100),
+            nullable=False,
+            comment="Disruption status (e.g., 'Good Service', 'Minor Delays', 'Severe Delays')",
+        ),
+        sa.Column(
+            "reason",
+            sa.String(length=1000),
+            nullable=True,
+            comment="Full disruption reason text (nullable for good service)",
+        ),
+        sa.Column(
+            "state_hash",
+            sa.String(length=64),
+            nullable=False,
+            comment="SHA256 hash of {line_id, status, reason} for deduplication",
+        ),
+        sa.Column(
+            "detected_at",
+            sa.DateTime(timezone=True),
+            nullable=False,
+            comment="When this state was detected by the alert service",
+        ),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_line_disruption_logs_line_detected", "line_disruption_state_logs", ["line_id", "detected_at"], unique=False
+    )
+    op.create_index(
+        op.f("ix_line_disruption_state_logs_detected_at"), "line_disruption_state_logs", ["detected_at"], unique=False
+    )
+    op.create_index(
+        op.f("ix_line_disruption_state_logs_line_id"), "line_disruption_state_logs", ["line_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_line_disruption_state_logs_state_hash"), "line_disruption_state_logs", ["state_hash"], unique=False
+    )
+    op.create_table(
+        "lines",
+        sa.Column("tfl_id", sa.String(length=50), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("mode", sa.String(length=50), server_default=sa.text("'tube'"), nullable=False),
+        sa.Column("route_variants", sa.JSON(), nullable=True),
+        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_lines_tfl_id"), "lines", ["tfl_id"], unique=True)
+    op.create_table(
+        "severity_codes",
+        sa.Column(
+            "mode_id",
+            sa.String(length=50),
+            nullable=False,
+            comment="Transport mode (e.g., 'tube', 'dlr', 'overground')",
+        ),
+        sa.Column("severity_level", sa.Integer(), nullable=False),
+        sa.Column("description", sa.String(length=255), nullable=False),
+        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("mode_id", "severity_level", name="uq_severity_code_mode_level"),
+    )
+    op.create_index(op.f("ix_severity_codes_mode_id"), "severity_codes", ["mode_id"], unique=False)
+    op.create_index(op.f("ix_severity_codes_severity_level"), "severity_codes", ["severity_level"], unique=False)
+    op.create_table(
+        "stations",
+        sa.Column("tfl_id", sa.String(length=50), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("latitude", sa.Float(), nullable=False),
+        sa.Column("longitude", sa.Float(), nullable=False),
+        sa.Column("lines", sa.JSON(), nullable=False),
+        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
+        sa.Column(
+            "hub_naptan_code",
+            sa.String(length=50),
+            nullable=True,
+            comment="TfL hub NaPTAN code for interchange stations",
+        ),
+        sa.Column(
+            "hub_common_name",
+            sa.String(length=255),
+            nullable=True,
+            comment="Common name for the hub (e.g., 'Seven Sisters')",
+        ),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_stations_hub_naptan_code"), "stations", ["hub_naptan_code"], unique=False)
+    op.create_index(op.f("ix_stations_name"), "stations", ["name"], unique=False)
+    op.create_index(op.f("ix_stations_tfl_id"), "stations", ["tfl_id"], unique=True)
+    op.create_table(
+        "stop_types",
+        sa.Column("type_name", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.String(length=500), nullable=True),
+        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_stop_types_type_name"), "stop_types", ["type_name"], unique=True)
+    op.create_table(
+        "users",
+        sa.Column("external_id", sa.String(length=255), nullable=False),
+        sa.Column("auth_provider", sa.String(length=50), server_default="auth0", nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index("ix_users_external_id_auth_provider", "users", ["external_id", "auth_provider"], unique=True)
+    op.create_table(
+        "admin_users",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("role", sa.Enum("admin", "superadmin", name="admin_role", create_constraint=True), nullable=False),
+        sa.Column("granted_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("granted_by", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["granted_by"], ["users.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_admin_users_user_id"), "admin_users", ["user_id"], unique=True)
+    op.create_table(
+        "email_addresses",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("email", sa.String(length=255), nullable=False),
+        sa.Column("verified", sa.Boolean(), nullable=False),
+        sa.Column("is_primary", sa.Boolean(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("email"),
+    )
+    op.create_index(op.f("ix_email_addresses_user_id"), "email_addresses", ["user_id"], unique=False)
+    op.create_index("ix_email_addresses_user_id_primary", "email_addresses", ["user_id", "is_primary"], unique=False)
+    op.create_table(
+        "phone_numbers",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("phone", sa.String(length=50), nullable=False),
+        sa.Column("verified", sa.Boolean(), nullable=False),
+        sa.Column("is_primary", sa.Boolean(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("phone"),
+    )
+    op.create_index(op.f("ix_phone_numbers_user_id"), "phone_numbers", ["user_id"], unique=False)
+    op.create_index("ix_phone_numbers_user_id_primary", "phone_numbers", ["user_id", "is_primary"], unique=False)
+    op.create_table(
+        "rate_limit_logs",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column(
+            "action_type",
+            sa.Enum("verify_code", "add_contact_failure", name="rate_limit_action", create_constraint=True),
+            nullable=False,
+        ),
+        sa.Column(
+            "resource_id", sa.String(length=255), nullable=False, comment="Contact ID or email/phone being verified"
+        ),
+        sa.Column("timestamp", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_rate_limit_logs_timestamp"), "rate_limit_logs", ["timestamp"], unique=False)
+    op.create_index(op.f("ix_rate_limit_logs_user_id"), "rate_limit_logs", ["user_id"], unique=False)
+    op.create_index("ix_rate_limit_resource_timestamp", "rate_limit_logs", ["resource_id", "timestamp"], unique=False)
+    op.create_index(
+        "ix_rate_limit_user_action_timestamp", "rate_limit_logs", ["user_id", "action_type", "timestamp"], unique=False
+    )
+    op.create_table(
+        "station_connections",
+        sa.Column("from_station_id", sa.UUID(), nullable=False),
+        sa.Column("to_station_id", sa.UUID(), nullable=False),
+        sa.Column("line_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["from_station_id"], ["stations.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["line_id"], ["lines.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["to_station_id"], ["stations.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index("ix_station_connections_from_station", "station_connections", ["from_station_id"], unique=False)
+    op.create_index("ix_station_connections_line", "station_connections", ["line_id"], unique=False)
+    op.create_index("ix_station_connections_to_station", "station_connections", ["to_station_id"], unique=False)
+    op.create_index(
+        "uq_station_connection_active",
+        "station_connections",
+        ["from_station_id", "to_station_id", "line_id"],
+        unique=True,
+        postgresql_where=sa.text("deleted_at IS NULL"),
+    )
+    op.create_table(
+        "station_disruptions",
+        sa.Column("station_id", sa.UUID(), nullable=False),
+        sa.Column(
+            "type",
+            sa.String(length=255),
+            nullable=True,
+            comment="Disruption type from TfL API (e.g., 'Information', 'Interchange Message')",
+        ),
+        sa.Column("description", sa.String(length=1000), nullable=False),
+        sa.Column(
+            "appearance",
+            sa.String(length=100),
+            nullable=True,
+            comment="Disruption appearance/status from TfL API (e.g., 'PlannedWork', 'RealTime')",
+        ),
+        sa.Column(
+            "tfl_id",
+            sa.String(length=100),
+            nullable=False,
+            comment="Hash-based identifier: atcoCode + fromDate + toDate + description",
+        ),
+        sa.Column(
+            "created_at_source",
+            sa.DateTime(timezone=True),
+            nullable=False,
+            comment="Start date from TfL API 'fromDate' field",
+        ),
+        sa.Column(
+            "end_date", sa.DateTime(timezone=True), nullable=True, comment="End date from TfL API 'toDate' field"
+        ),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["station_id"], ["stations.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_station_disruptions_station_id"), "station_disruptions", ["station_id"], unique=False)
+    op.create_index(op.f("ix_station_disruptions_tfl_id"), "station_disruptions", ["tfl_id"], unique=False)
+    op.create_table(
+        "user_routes",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("active", sa.Boolean(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column(
+            "timezone",
+            sa.String(length=64),
+            server_default=sa.text("'Europe/London'"),
+            nullable=False,
+            comment="IANA timezone for schedule interpretation",
+        ),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="RESTRICT"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_user_routes_user_id"), "user_routes", ["user_id"], unique=False)
+    op.create_table(
+        "verification_codes",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column(
+            "contact_id", sa.UUID(), nullable=False, comment="UUID of the EmailAddress or PhoneNumber being verified"
+        ),
+        sa.Column("code", sa.String(length=6), nullable=False),
+        sa.Column("type", sa.Enum("email", "sms", name="verification_type", create_constraint=True), nullable=False),
+        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("used", sa.Boolean(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_verification_codes_contact_id"), "verification_codes", ["contact_id"], unique=False)
+    op.create_index(
+        "ix_verification_codes_user_code_type", "verification_codes", ["user_id", "code", "type"], unique=False
+    )
+    op.create_index(op.f("ix_verification_codes_user_id"), "verification_codes", ["user_id"], unique=False)
+    op.create_table(
+        "notification_logs",
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("route_id", sa.UUID(), nullable=False),
+        sa.Column("sent_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column(
+            "method", sa.Enum("email", "sms", name="notification_method", create_constraint=True), nullable=False
+        ),
+        sa.Column(
+            "status",
+            sa.Enum("sent", "failed", "pending", name="notification_status", create_constraint=True),
+            nullable=False,
+        ),
+        sa.Column("error_message", sa.Text(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_notification_logs_route_id"), "notification_logs", ["route_id"], unique=False)
+    op.create_index("ix_notification_logs_route_sent", "notification_logs", ["route_id", "sent_at"], unique=False)
+    op.create_index(op.f("ix_notification_logs_sent_at"), "notification_logs", ["sent_at"], unique=False)
+    op.create_index(op.f("ix_notification_logs_user_id"), "notification_logs", ["user_id"], unique=False)
+    op.create_index("ix_notification_logs_user_sent", "notification_logs", ["user_id", "sent_at"], unique=False)
+    op.create_table(
+        "notification_preferences",
+        sa.Column("route_id", sa.UUID(), nullable=False),
+        sa.Column(
+            "method", sa.Enum("email", "sms", name="notification_method", create_constraint=True), nullable=False
+        ),
+        sa.Column("target_email_id", sa.UUID(), nullable=True),
+        sa.Column("target_phone_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "(target_email_id IS NOT NULL AND target_phone_id IS NULL) OR (target_email_id IS NULL AND target_phone_id IS NOT NULL)",
+            name="ck_notification_preference_target",
+        ),
+        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="RESTRICT"),
+        sa.ForeignKeyConstraint(["target_email_id"], ["email_addresses.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["target_phone_id"], ["phone_numbers.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index("ix_notification_preferences_route", "notification_preferences", ["route_id"], unique=False)
+    op.create_index(
+        op.f("ix_notification_preferences_route_id"), "notification_preferences", ["route_id"], unique=False
+    )
+    op.create_table(
+        "user_route_schedules",
+        sa.Column("route_id", sa.UUID(), nullable=False),
+        sa.Column("days_of_week", sa.JSON(), nullable=False),
+        sa.Column("start_time", sa.Time(), nullable=False),
+        sa.Column("end_time", sa.Time(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="RESTRICT"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_user_route_schedules_route_id"), "user_route_schedules", ["route_id"], unique=False)
+    op.create_table(
+        "user_route_segments",
+        sa.Column("route_id", sa.UUID(), nullable=False),
+        sa.Column("sequence", sa.Integer(), nullable=False),
+        sa.Column("station_id", sa.UUID(), nullable=False),
+        sa.Column("line_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["line_id"], ["lines.id"], ondelete="RESTRICT"),
+        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="RESTRICT"),
+        sa.ForeignKeyConstraint(["station_id"], ["stations.id"], ondelete="RESTRICT"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_user_route_segments_route_id"), "user_route_segments", ["route_id"], unique=False)
+    op.create_index(
+        "ix_user_route_segments_route_sequence", "user_route_segments", ["route_id", "sequence"], unique=False
+    )
+    op.create_index(
+        "uq_route_segment_sequence_active",
+        "user_route_segments",
+        ["route_id", "sequence"],
+        unique=True,
+        postgresql_where=sa.text("deleted_at IS NULL"),
+    )
+    op.create_table(
+        "user_route_station_index",
+        sa.Column("route_id", sa.UUID(), nullable=False, comment="User's route ID"),
+        sa.Column(
+            "line_tfl_id", sa.String(length=50), nullable=False, comment="TfL line ID (e.g., 'piccadilly', 'northern')"
+        ),
+        sa.Column(
+            "station_naptan", sa.String(length=50), nullable=False, comment="Station NaPTAN code (e.g., '940GZZLUKSX')"
+        ),
+        sa.Column(
+            "line_data_version",
+            sa.DateTime(timezone=True),
+            nullable=False,
+            comment="Copy of Line.last_updated for staleness detection",
+        ),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="RESTRICT"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_user_route_station_index_line_data_version", "user_route_station_index", ["line_data_version"], unique=False
+    )
+    op.create_index(
+        "ix_user_route_station_index_line_station",
+        "user_route_station_index",
+        ["line_tfl_id", "station_naptan"],
+        unique=False,
+    )
+    op.create_index("ix_user_route_station_index_route", "user_route_station_index", ["route_id"], unique=False)
+
+    # Populate alert_disabled_severities with Good Service (severity 10) for all modes
+    # This ensures "Good Service" status does not trigger alerts
+    tfl_modes = [
+        "tube",
+        "dlr",
+        "overground",
+        "elizabeth-line",
+        "tram",
+        "tfl-rail",
+        "cable-car",
+        "river-bus",
+        "national-rail",
+    ]
+
+    for mode in tfl_modes:
+        op.execute(
+            f"""
+            INSERT INTO alert_disabled_severities (id, mode_id, severity_level, created_at, updated_at)
+            VALUES (gen_random_uuid(), '{mode}', 10, now(), now())
+            """
+        )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    msg = "Squashed migration has no downgrade path"
+    raise NotImplementedError(msg)
