--- backend/alembic/versions/49b26f7f0a3a_add_rate_limit_log_model.py
@@ -1,60 +1,0 @@
-"""Add rate limit log model
-
-Revision ID: 49b26f7f0a3a
-Revises: 883740308b56
-Create Date: 2025-11-02 18:05:41.261872
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "49b26f7f0a3a"
-down_revision: str | Sequence[str] | None = "883740308b56"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "rate_limit_logs",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column(
-            "action_type",
-            sa.Enum("verify_code", "add_contact_failure", name="rate_limit_action", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column(
-            "resource_id", sa.String(length=255), nullable=False, comment="Contact ID or email/phone being verified"
-        ),
-        sa.Column("timestamp", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_rate_limit_logs_timestamp"), "rate_limit_logs", ["timestamp"], unique=False)
-    op.create_index(op.f("ix_rate_limit_logs_user_id"), "rate_limit_logs", ["user_id"], unique=False)
-    op.create_index("ix_rate_limit_resource_timestamp", "rate_limit_logs", ["resource_id", "timestamp"], unique=False)
-    op.create_index(
-        "ix_rate_limit_user_action_timestamp", "rate_limit_logs", ["user_id", "action_type", "timestamp"], unique=False
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index("ix_rate_limit_user_action_timestamp", table_name="rate_limit_logs")
-    op.drop_index("ix_rate_limit_resource_timestamp", table_name="rate_limit_logs")
-    op.drop_index(op.f("ix_rate_limit_logs_user_id"), table_name="rate_limit_logs")
-    op.drop_index(op.f("ix_rate_limit_logs_timestamp"), table_name="rate_limit_logs")
-    op.drop_table("rate_limit_logs")
-    # ### end Alembic commands ###
