--- backend/alembic/versions/883740308b56_initial_schema_with_all_models.py
@@ -1,335 +1,0 @@
-"""Initial schema with all models
-
-Revision ID: 883740308b56
-Revises:
-Create Date: 2025-11-01 20:47:37.645102
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "883740308b56"
-down_revision: str | Sequence[str] | None = None
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "lines",
-        sa.Column("tfl_id", sa.String(length=50), nullable=False),
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column("color", sa.String(length=7), nullable=False),
-        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_lines_tfl_id"), "lines", ["tfl_id"], unique=True)
-    op.create_table(
-        "stations",
-        sa.Column("tfl_id", sa.String(length=50), nullable=False),
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column("latitude", sa.Float(), nullable=False),
-        sa.Column("longitude", sa.Float(), nullable=False),
-        sa.Column("lines", sa.JSON(), nullable=False),
-        sa.Column("last_updated", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_stations_name"), "stations", ["name"], unique=False)
-    op.create_index(op.f("ix_stations_tfl_id"), "stations", ["tfl_id"], unique=True)
-    op.create_table(
-        "users",
-        sa.Column("external_id", sa.String(length=255), nullable=False),
-        sa.Column("auth_provider", sa.String(length=50), nullable=False, server_default="auth0"),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        "ix_users_external_id_auth_provider",
-        "users",
-        ["external_id", "auth_provider"],
-        unique=True,
-    )
-    op.create_table(
-        "admin_users",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column(
-            "role",
-            sa.Enum("admin", "superadmin", name="admin_role", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column("granted_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("granted_by", sa.UUID(), nullable=True),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["granted_by"], ["users.id"], ondelete="SET NULL"),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_admin_users_user_id"), "admin_users", ["user_id"], unique=True)
-    op.create_table(
-        "email_addresses",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("email", sa.String(length=255), nullable=False),
-        sa.Column("verified", sa.Boolean(), nullable=False),
-        sa.Column("is_primary", sa.Boolean(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("email"),
-    )
-    op.create_index(op.f("ix_email_addresses_user_id"), "email_addresses", ["user_id"], unique=False)
-    op.create_index(
-        "ix_email_addresses_user_id_primary",
-        "email_addresses",
-        ["user_id", "is_primary"],
-        unique=False,
-    )
-    op.create_table(
-        "phone_numbers",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("phone", sa.String(length=50), nullable=False),
-        sa.Column("verified", sa.Boolean(), nullable=False),
-        sa.Column("is_primary", sa.Boolean(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("phone"),
-    )
-    op.create_index(op.f("ix_phone_numbers_user_id"), "phone_numbers", ["user_id"], unique=False)
-    op.create_index("ix_phone_numbers_user_id_primary", "phone_numbers", ["user_id", "is_primary"], unique=False)
-    op.create_table(
-        "user_routes",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column("active", sa.Boolean(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_user_routes_user_id"), "user_routes", ["user_id"], unique=False)
-    op.create_table(
-        "station_connections",
-        sa.Column("from_station_id", sa.UUID(), nullable=False),
-        sa.Column("to_station_id", sa.UUID(), nullable=False),
-        sa.Column("line_id", sa.UUID(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["from_station_id"], ["stations.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["line_id"], ["lines.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["to_station_id"], ["stations.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("from_station_id", "to_station_id", "line_id", name="uq_station_connection"),
-    )
-    op.create_index(
-        "ix_station_connections_from_station",
-        "station_connections",
-        ["from_station_id"],
-        unique=False,
-    )
-    op.create_index("ix_station_connections_line", "station_connections", ["line_id"], unique=False)
-    op.create_index("ix_station_connections_to_station", "station_connections", ["to_station_id"], unique=False)
-    op.create_table(
-        "verification_codes",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("code", sa.String(length=6), nullable=False),
-        sa.Column(
-            "type",
-            sa.Enum("email", "sms", name="verification_type", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("used", sa.Boolean(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        "ix_verification_codes_user_code_type",
-        "verification_codes",
-        ["user_id", "code", "type"],
-        unique=False,
-    )
-    op.create_index(op.f("ix_verification_codes_user_id"), "verification_codes", ["user_id"], unique=False)
-    op.create_table(
-        "notification_logs",
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("route_id", sa.UUID(), nullable=False),
-        sa.Column("sent_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column(
-            "method",
-            sa.Enum("email", "sms", name="notification_method", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column(
-            "status",
-            sa.Enum("sent", "failed", "pending", name="notification_status", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column("error_message", sa.Text(), nullable=True),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_notification_logs_route_id"), "notification_logs", ["route_id"], unique=False)
-    op.create_index(
-        "ix_notification_logs_route_sent",
-        "notification_logs",
-        ["route_id", "sent_at"],
-        unique=False,
-    )
-    op.create_index(op.f("ix_notification_logs_sent_at"), "notification_logs", ["sent_at"], unique=False)
-    op.create_index(op.f("ix_notification_logs_user_id"), "notification_logs", ["user_id"], unique=False)
-    op.create_index("ix_notification_logs_user_sent", "notification_logs", ["user_id", "sent_at"], unique=False)
-    op.create_table(
-        "notification_preferences",
-        sa.Column("route_id", sa.UUID(), nullable=False),
-        sa.Column(
-            "method",
-            sa.Enum("email", "sms", name="notification_method", create_constraint=True),
-            nullable=False,
-        ),
-        sa.Column("target_email_id", sa.UUID(), nullable=True),
-        sa.Column("target_phone_id", sa.UUID(), nullable=True),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "(target_email_id IS NOT NULL AND target_phone_id IS NULL) OR (target_email_id IS NULL AND target_phone_id IS NOT NULL)",
-            name="ck_notification_preference_target",
-        ),
-        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["target_email_id"], ["email_addresses.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["target_phone_id"], ["phone_numbers.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index("ix_notification_preferences_route", "notification_preferences", ["route_id"], unique=False)
-    op.create_index(
-        op.f("ix_notification_preferences_route_id"),
-        "notification_preferences",
-        ["route_id"],
-        unique=False,
-    )
-    op.create_table(
-        "user_route_schedules",
-        sa.Column("route_id", sa.UUID(), nullable=False),
-        sa.Column("days_of_week", sa.JSON(), nullable=False),
-        sa.Column("start_time", sa.Time(), nullable=False),
-        sa.Column("end_time", sa.Time(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_user_route_schedules_route_id"), "user_route_schedules", ["route_id"], unique=False)
-    op.create_table(
-        "user_route_segments",
-        sa.Column("route_id", sa.UUID(), nullable=False),
-        sa.Column("sequence", sa.Integer(), nullable=False),
-        sa.Column("station_id", sa.UUID(), nullable=False),
-        sa.Column("line_id", sa.UUID(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["line_id"], ["lines.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["route_id"], ["user_routes.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["station_id"], ["stations.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("route_id", "sequence", name="uq_route_segment_sequence"),
-    )
-    op.create_index(op.f("ix_user_route_segments_route_id"), "user_route_segments", ["route_id"], unique=False)
-    op.create_index(
-        "ix_user_route_segments_route_sequence", "user_route_segments", ["route_id", "sequence"], unique=False
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index("ix_user_route_segments_route_sequence", table_name="user_route_segments")
-    op.drop_index(op.f("ix_user_route_segments_route_id"), table_name="user_route_segments")
-    op.drop_table("user_route_segments")
-    op.drop_index(op.f("ix_user_route_schedules_route_id"), table_name="user_route_schedules")
-    op.drop_table("user_route_schedules")
-    op.drop_index(op.f("ix_notification_preferences_route_id"), table_name="notification_preferences")
-    op.drop_index("ix_notification_preferences_route", table_name="notification_preferences")
-    op.drop_table("notification_preferences")
-    op.drop_index("ix_notification_logs_user_sent", table_name="notification_logs")
-    op.drop_index(op.f("ix_notification_logs_user_id"), table_name="notification_logs")
-    op.drop_index(op.f("ix_notification_logs_sent_at"), table_name="notification_logs")
-    op.drop_index("ix_notification_logs_route_sent", table_name="notification_logs")
-    op.drop_index(op.f("ix_notification_logs_route_id"), table_name="notification_logs")
-    op.drop_table("notification_logs")
-    op.drop_index(op.f("ix_verification_codes_user_id"), table_name="verification_codes")
-    op.drop_index("ix_verification_codes_user_code_type", table_name="verification_codes")
-    op.drop_table("verification_codes")
-    op.drop_index("ix_station_connections_to_station", table_name="station_connections")
-    op.drop_index("ix_station_connections_line", table_name="station_connections")
-    op.drop_index("ix_station_connections_from_station", table_name="station_connections")
-    op.drop_table("station_connections")
-    op.drop_index(op.f("ix_user_routes_user_id"), table_name="user_routes")
-    op.drop_table("user_routes")
-    op.drop_index("ix_phone_numbers_user_id_primary", table_name="phone_numbers")
-    op.drop_index(op.f("ix_phone_numbers_user_id"), table_name="phone_numbers")
-    op.drop_table("phone_numbers")
-    op.drop_index("ix_email_addresses_user_id_primary", table_name="email_addresses")
-    op.drop_index(op.f("ix_email_addresses_user_id"), table_name="email_addresses")
-    op.drop_table("email_addresses")
-    op.drop_index(op.f("ix_admin_users_user_id"), table_name="admin_users")
-    op.drop_table("admin_users")
-    op.drop_index("ix_users_external_id_auth_provider", table_name="users")
-    op.drop_table("users")
-    op.drop_index(op.f("ix_stations_tfl_id"), table_name="stations")
-    op.drop_index(op.f("ix_stations_name"), table_name="stations")
-    op.drop_table("stations")
-    op.drop_index(op.f("ix_lines_tfl_id"), table_name="lines")
-    op.drop_table("lines")
-    # ### end Alembic commands ###
-
-    # Drop enum types
-    op.execute("DROP TYPE IF EXISTS admin_role CASCADE")
-    op.execute("DROP TYPE IF EXISTS verification_type CASCADE")
-    op.execute("DROP TYPE IF EXISTS notification_method CASCADE")
-    op.execute("DROP TYPE IF EXISTS notification_status CASCADE")
