--- backend/alembic/versions/e9121fedff17_add_contact_id_to_verification_codes.py
@@ -1,39 +1,0 @@
-"""Add contact_id to verification_codes
-
-Revision ID: e9121fedff17
-Revises: 49b26f7f0a3a
-Create Date: 2025-11-02 20:10:49.825155
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "e9121fedff17"
-down_revision: str | Sequence[str] | None = "49b26f7f0a3a"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "verification_codes",
-        sa.Column(
-            "contact_id", sa.UUID(), nullable=False, comment="UUID of the EmailAddress or PhoneNumber being verified"
-        ),
-    )
-    op.create_index(op.f("ix_verification_codes_contact_id"), "verification_codes", ["contact_id"], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_verification_codes_contact_id"), table_name="verification_codes")
-    op.drop_column("verification_codes", "contact_id")
-    # ### end Alembic commands ###
