--- backend/alembic/versions/f8838ec5262d_remove_redundant_station_id_index_from_.py
@@ -1,84 +1,0 @@
-"""Remove redundant station_id index from station_disruptions
-
-Revision ID: f8838ec5262d
-Revises: a1c51fa40ade
-Create Date: 2025-11-05 12:15:40.404394
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "f8838ec5262d"
-down_revision: str | Sequence[str] | None = "a1c51fa40ade"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f("disruption_categories_category_name_key"), "disruption_categories", type_="unique")
-    op.drop_index(op.f("ix_disruption_categories_category_name"), table_name="disruption_categories")
-    op.create_index(
-        op.f("ix_disruption_categories_category_name"), "disruption_categories", ["category_name"], unique=True
-    )
-    op.alter_column(
-        "user_routes",
-        "timezone",
-        existing_type=sa.VARCHAR(length=64),
-        comment="IANA timezone for schedule interpretation",
-        existing_comment="IANA timezone for schedule interpretation (e.g., Europe/London). Schedule times are naive and interpreted in this timezone.",
-        existing_nullable=False,
-        existing_server_default=sa.text("'Europe/London'::character varying"),
-    )
-    op.drop_constraint(op.f("severity_codes_severity_level_key"), "severity_codes", type_="unique")
-    op.drop_index(op.f("ix_severity_codes_severity_level"), table_name="severity_codes")
-    op.create_index(op.f("ix_severity_codes_severity_level"), "severity_codes", ["severity_level"], unique=True)
-    op.drop_index(op.f("ix_station_disruptions_station"), table_name="station_disruptions")
-    op.drop_constraint(op.f("stop_types_type_name_key"), "stop_types", type_="unique")
-    op.drop_index(op.f("ix_stop_types_type_name"), table_name="stop_types")
-    op.create_index(op.f("ix_stop_types_type_name"), "stop_types", ["type_name"], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_stop_types_type_name"), table_name="stop_types")
-    op.create_index(op.f("ix_stop_types_type_name"), "stop_types", ["type_name"], unique=False)
-    op.create_unique_constraint(
-        op.f("stop_types_type_name_key"), "stop_types", ["type_name"], postgresql_nulls_not_distinct=False
-    )
-    op.create_index(op.f("ix_station_disruptions_station"), "station_disruptions", ["station_id"], unique=False)
-    op.drop_index(op.f("ix_severity_codes_severity_level"), table_name="severity_codes")
-    op.create_index(op.f("ix_severity_codes_severity_level"), "severity_codes", ["severity_level"], unique=False)
-    op.create_unique_constraint(
-        op.f("severity_codes_severity_level_key"),
-        "severity_codes",
-        ["severity_level"],
-        postgresql_nulls_not_distinct=False,
-    )
-    op.alter_column(
-        "user_routes",
-        "timezone",
-        existing_type=sa.VARCHAR(length=64),
-        comment="IANA timezone for schedule interpretation (e.g., Europe/London). Schedule times are naive and interpreted in this timezone.",
-        existing_comment="IANA timezone for schedule interpretation",
-        existing_nullable=False,
-        existing_server_default=sa.text("'Europe/London'::character varying"),
-    )
-    op.drop_index(op.f("ix_disruption_categories_category_name"), table_name="disruption_categories")
-    op.create_index(
-        op.f("ix_disruption_categories_category_name"), "disruption_categories", ["category_name"], unique=False
-    )
-    op.create_unique_constraint(
-        op.f("disruption_categories_category_name_key"),
-        "disruption_categories",
-        ["category_name"],
-        postgresql_nulls_not_distinct=False,
-    )
-    # ### end Alembic commands ###
